Move the scraper up in the path
The namespace of videodata/videodata was an artifact of how I initially started
the project. Before this goes too far I've pulled this up one directory and
made a simple setup.py to make it easier to silo tests.

This commit also move the extract_speakers method into the youtube module
because this method is VERY specific to the layout of a youtube description.
It's possible for this to change from year to year.
