+++ tox.ini
@@ -14,4 +14,4 @@ basepython = python3.4
 deps = -rtest_requirements.txt
 commands =
   pycodestyle ./
-  py.test ./videodata/tests
+  py.test ./tests
